fix(sync): fresh envelope + fresh messageId per retry attempt (rc.9 PR-E codex follow-up #8 + design simplification)
Codex flagged a final correctness issue on 4b3d9b08:

  > Hoisting buildSyncRequest(...) out of the send retry loop
  > reuses the same signed sync envelope across every transport
  > retry. For private context graphs that envelope carries
  > issuedAtMs/requestId, and the current retry budget can exceed
  > the 90s auth freshness window (SYNC_PAGE_TIMEOUT_MS is 45s,
  > SYNC_PAGE_RETRY_ATTEMPTS is 3, SYNC_AUTH_MAX_AGE_MS is 90s).

This is the SEVENTH bug in a chain (follow-ups #1, #2, #4, #5, #6,
#7, #8). Each intermediate "stable messageId" design satisfied one
finding while introducing another. Stepping back:

The root issue is that sync's app-layer envelope already enforces
identity / freshness / replay (issuedAtMs TTL + per-requestId
replay cache + signature). Trying to also enforce "logical
request identity" at the substrate layer (stable messageId) is
fundamentally incompatible — the two layers measure freshness on
different clocks and the substrate's 24h-default outbox window
makes it impossible to keep envelope-bound metadata in sync with
substrate-bound metadata.

## Final design: fresh-per-attempt for everything

  - `sync-transport.ts`: `withRetry` inner closure calls
    `requestFactory()` (fresh envelope) AND `randomUUID()` (fresh
    messageId) on every attempt. Removed the `payload` /
    `messageId` plumbing introduced for the stable-id designs.

  - `page-fetch.ts`: passes `requestFactory: () => buildSyncRequest(...)`
    to `sendSyncRequest`. Reverts to the pre-PR-E topology for
    page iteration. Drops `computeSyncPageMessageId`, the
    per-fetchSyncPages `runNonce`, the RESPONSE_GONE recovery
    block, and the explicit per-page build-retry wrapper (the
    build is back inside `sendSyncRequest`'s `withRetry`, so
    transient build failures retry as before).

  - `dkg-agent.ts` send adapter: simple `messenger.sendReliable`
    wrapper, passes the fresh per-attempt `messageId` straight
    through. ALSO passes `maxAgeMs: SYNC_AUTH_MAX_AGE_MS - 5_000`
    so the substrate's outbox drops sync entries within ~85s of
    first failure — comfortably before any envelope they hold
    could be denied as stale, and protecting against orphaned
    outbox entries hanging around for 24h doing redundant work.

## Trade-offs

(+) Correct under all timing scenarios. Each attempt's envelope
    is fresh-at-send-time; each attempt's messageId is unique so
    no cached response (success or denial) ever replays past its
    intended attempt.

(+) No RESPONSE_GONE recovery needed — RESPONSE_GONE only fires
    on duplicate-receive for the SAME messageId, and we never
    send the same messageId twice.

(-) No sender-side dedup of network-retry storms within a single
    page-fetch. If a transport hiccup causes 3 retries and all 3
    happen to reach the receiver, the responder runs the SPARQL
    page query 3 times. Bounded by syncPageRetryAttempts; sync
    queries are app-layer idempotent so this is wasted-work, not
    a correctness concern. (The original codex bug #1 wanted
    stable messageIds for exactly this dedup; given the cascade,
    the perf win wasn't worth the correctness invariants.)

## Tests

Deleted `test/sync-page-messageid-stability.test.ts` (its
assertions were all about stable-messageId behaviour that no
longer exists). Replaced with `test/sync-fresh-per-attempt.test.ts`
(4 tests) that pin the new invariant. Each test cites the codex
finding it regresses:

  - `passes a non-empty messageId on every send invocation` —
    follow-up #1: the page-fetch send contract guarantees a
    populated id (the substrate adapter then forwards or rotates
    it; the FRESH-per-attempt behaviour itself lives in
    sync-transport.ts and is covered by the next test).
  - `mints a DIFFERENT messageId on every withRetry attempt for
    the same page` — follow-up #8: forces 3 retries, asserts all
    3 ids are distinct. Prevents reintroduction of stable-id
    optimisation that the cascade showed is unsafe.
  - `rebuilds the request envelope on every withRetry attempt
    (fresh issuedAtMs/requestId)` — follow-up #5 + #8: 3 retries,
    asserts buildSyncRequest ran 3x AND that each retry's send
    received the bytes from THAT attempt's build (not a stale
    earlier one).
  - `retries transient envelope-build failures (build inside
    withRetry)` — follow-up #7: 2 build throws + 1 success, send
    fires exactly once. Confirms the build-time retry coverage
    that was implicit in the pre-PR-E topology is preserved.

Full sync regression suite green: 35/35 across
sync-fresh-per-attempt (4 NEW), dkg-agent-diagnostics (16),
sync-responder-per-cgid-meta (5), sync-on-connect-retry (10).
Agent typecheck clean.
